[rotorcraft] care free mode
- still not implemented for stabilization_euler
- move read_rc_setpoint_* functions to c file
- add GUIDANCE_H_MODE_CARE_FREE mode
- add stabilization float_euler subsystem
- use normal stabilization subsystem in NPS instead of hardcoded euler
