+++ sw/airborne/firmwares/rotorcraft/autopilot.h
@@ -50,7 +50,7 @@
 #define AP_MODE_HOVER_Z_HOLD      11
 #define AP_MODE_NAV               12
 #define AP_MODE_RC_DIRECT         13	// Safety Pilot Direct Commands for helicopter direct control: appropriately chosen as mode "13"
-#define AP_MODE_CARE_FREE         14
+#define AP_MODE_CARE_FREE_DIRECT  14
 
 extern uint8_t autopilot_mode;
 extern uint8_t autopilot_mode_auto2;
