+++ sw/airborne/firmwares/rotorcraft/stabilization/stabilization_attitude.h
@@ -25,8 +25,6 @@
 
 #include STABILIZATION_ATTITUDE_TYPE_H
 
-extern float care_free_heading;
-
 extern void stabilization_attitude_init(void);
 extern void stabilization_attitude_read_rc(bool_t in_flight);
 extern void stabilization_attitude_enter(void);
